1단계. StringCalculatorInputValidator 를 OperandValidator 로 변경
